added a working personal page
