+++ dbh.inc.php
@@ -1,5 +1,4 @@
 <?php
-
     $dbhost = "localhost";
     $dbuser = "root";
     $dbpassword = "";
