+++ functions.inc.php
@@ -1,8 +1,4 @@
 <?php
-    ini_set('display_errors', 1);
-    ini_set('display_startup_errors', 1);
-    error_reporting(E_ALL);
-
     require_once 'dbh.inc.php';
     function emptyInputSignup($name, $email, $password, $password2) {
         $result = false;
@@ -54,6 +50,29 @@
         }
 
         mysqli_stmt_bind_param($stmt, "s", $email);
+        mysqli_stmt_execute($stmt);
+
+        $resultData = mysqli_stmt_get_result($stmt);
+
+        if ($row = mysqli_fetch_assoc($resultData)) {
+            return $row;
+        } else {
+            $result = false;
+        }
+
+        mysqli_stmt_close($stmt);
+        return $result;
+    }
+
+    function dataExists($conn, $userID) {
+        $sql = "SELECT * FROM personal WHERE userID = ?;";
+        $stmt = mysqli_stmt_init($conn);
+        if (!mysqli_stmt_prepare($stmt, $sql)) {
+            header("location: login.php?error=stmtfailed");
+            exit();
+        }
+
+        mysqli_stmt_bind_param($stmt, "s", $userID);
         mysqli_stmt_execute($stmt);
 
         $resultData = mysqli_stmt_get_result($stmt);
@@ -130,6 +149,16 @@
             $_SESSION["id"] = $emailExists["id"];
             $_SESSION["email"] = $emailExists["email"];
             $_SESSION["name"] = $emailExists["name"];
+
+            $dataExists = dataExists($conn, $_SESSION["id"]);
+
+            $_SESSION["bio"] = $dataExists["bio"];
+            $_SESSION["favChin"] = $dataExists["favChin"];
+            $_SESSION["owner"] = $dataExists["owner"];
+            $_SESSION["birthDate"] = $dataExists["birthDate"];
+            $_SESSION["dailyTime"] = $dataExists["dailyTime"];
+            $_SESSION["loveTime"] = $dataExists["loveTime"];
+
             header("location: elerheto.php");
             exit();
         }
